+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.2.1] - 2023-01-31
+
+### Added
+
+- `IStateMachine` interface.
+
+### Changed
+
+- `StateMachine` now implements `IStateMachine`.
+- StateMachines now reset their state holders when stopped.
+
 ## [0.2.0] - 2022-04-15
 
 ### Added